Reduce length of Firestore.class
Now some functions used to communicate with the server are defined
in the class that use them.
