+++ app/src/androidTest/java/ch/epfl/sweng/studyup/NavigationTest.java
@@ -19,10 +19,19 @@ import ch.epfl.sweng.studyup.map.MapsActivity;
 import ch.epfl.sweng.studyup.player.HomeActivity;
 import ch.epfl.sweng.studyup.player.QuestsActivityStudent;
 import ch.epfl.sweng.studyup.player.ScheduleActivityStudent;
+import ch.epfl.sweng.studyup.settings.SettingsActivity;
 
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.click;
+import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.intent.Intents.intended;
 import static android.support.test.espresso.intent.matcher.IntentMatchers.hasComponent;
+import static android.support.test.espresso.matcher.RootMatchers.withDecorView;
+import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static android.support.test.espresso.matcher.ViewMatchers.withText;
 import static ch.epfl.sweng.studyup.utils.GlobalAccessVariables.MOCK_ENABLED;
+import static org.hamcrest.core.IsNot.not;
 
 @RunWith(AndroidJUnit4.class)
 public class NavigationTest {
@@ -65,13 +74,13 @@ public class NavigationTest {
         intended(hasComponent(ScheduleActivityStudent.class.getName()));
     }
 
-    /*@Test
+    @Test
     public void navigationTopTest() {
         onView(withId(R.id.top_navigation_infos)).perform(click());
         onView(withText("Infos are coming soon")).inRoot(withDecorView(not(mActivityRule.getActivity().getWindow().getDecorView()))).check(matches(isDisplayed()));
         onView(withId(R.id.top_navigation_settings)).perform(click());
         intended(hasComponent(SettingsActivity.class.getName()));
-    }*/
+    }
 
 }
